Transaction types api

+++ src/transaction-types/transaction-types.service.ts
@@ -4,10 +4,12 @@ import {
   CreateOptions,
   DestroyOptions,
   FindOptions,
-  RestoreOptions,
   UpdateOptions,
 } from 'sequelize';
+import { UtilsProvider } from 'src/common/utils/utils.provider';
+import { Transaction } from 'src/transactions/entities/transaction.entity';
 import { CreateTransactionTypeDto } from './dto/create-transaction-type.dto';
+import { FindAllTransactionTypesQuery } from './dto/find-all-transaction-types.dto';
 import { UpdateTransactionTypeDto } from './dto/update-transaction-type.dto';
 import { TransactionType } from './entities/transaction-type.entity';
 
@@ -16,6 +18,9 @@ export class TransactionTypesService {
   constructor(
     @InjectModel(TransactionType)
     private transactionTypeModel: typeof TransactionType,
+    @InjectModel(Transaction)
+    private transactionModel: typeof Transaction,
+    private readonly utilsProvider: UtilsProvider,
   ) {}
 
   async create(
@@ -28,8 +33,57 @@ export class TransactionTypesService {
     );
   }
 
-  async findAll(options?: FindOptions): Promise<TransactionType[]> {
-    return await this.transactionTypeModel.findAll(options);
+  async findAllWithCount(
+    query: FindAllTransactionTypesQuery,
+    options?: FindOptions,
+  ): Promise<{ count: number; rows: TransactionType[] }> {
+    const filterClauses =
+      this.utilsProvider.queryBuilder.whereClausesFromFilters<TransactionType>(
+        'TransactionType',
+        query,
+        [
+          { field: 'id', matchType: 'exact' },
+          { field: 'name', matchType: 'fuzzy' },
+          { field: 'isDeduction', matchType: 'yes-no-boolean' },
+          { field: 'description', matchType: 'fuzzy' },
+          { field: 'status', matchType: 'exact' },
+          { field: 'createdAt', matchType: 'date' },
+          { field: 'updatedAt', matchType: 'date' },
+        ],
+      );
+
+    const searchClauses =
+      this.utilsProvider.queryBuilder.whereClausesFromSearch<TransactionType>(
+        'TransactionType',
+        query.search,
+        [
+          { field: 'name', matchType: 'fuzzy' },
+          { field: 'description', matchType: 'fuzzy' },
+        ],
+      );
+
+    const where = this.utilsProvider.queryBuilder.joinWhereClauses('and', [
+      filterClauses,
+      searchClauses,
+    ]);
+
+    const order = this.utilsProvider.queryBuilder.orderByField<TransactionType>(
+      ['createdAt', 'desc'],
+      [query.sortBy as keyof TransactionType, query.sortOrder],
+    );
+
+    const [offset, limit] = this.utilsProvider.queryBuilder.pagination(
+      query.page,
+      query.perPage,
+    );
+
+    return await this.transactionTypeModel.findAndCountAll({
+      where,
+      order,
+      offset,
+      limit,
+      ...(options || {}),
+    });
   }
 
   async findById(id: number, options?: FindOptions): Promise<TransactionType> {
@@ -49,22 +103,23 @@ export class TransactionTypesService {
   }
 
   async remove(id: number, options?: DestroyOptions): Promise<TransactionType> {
+    const transactionsWithType = await this.transactionModel.findOne({
+      where: { transactionTypeId: id },
+    });
+
+    if (transactionsWithType) {
+      throw new Error(
+        'Cannot  delete Transaction Type because it has Transactions',
+      );
+    }
+
     const transactionType = await this.findById(id);
+
     await this.transactionTypeModel.destroy({
       where: { id },
       ...(options || {}),
     });
-    return transactionType;
-  }
 
-  async restore(
-    id: number,
-    options?: RestoreOptions,
-  ): Promise<TransactionType> {
-    await this.transactionTypeModel.restore({
-      where: { id },
-      ...(options || {}),
-    });
-    return await this.findById(id);
+    return transactionType;
   }
 }
